+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 - `std` and `const` features for more control
 - `Date::epoch` function to provide the Epoch date for unconst envs (like wasm)
+- `Date::from_d_past_epoch` and `Date::from_d_before_epoch`
 
 # `0.3.0`